// app/interfaces/Agency.tsx
export default interface Agency {
	name: string;
	state: string;
	state_code: string;
	type: string;
	population: string;
	website: string;
	total_schools: string;
	total_students: string;
	mailing_address: string;
	grade_span: string;
	locale: string;
	csa_cbsa: string;
	domain_name: string;
	physical_address: string;
	phone: string;
	status: string;
	student_teacher_ratio: string;
	supervisory_union: string;
	county: string;
	created_at: string;
	updated_at: string;
	id: string;
}